screen-cast: Generalize and move pipewire code
So that it can be reused outside of the screen-cast code.
